fix solve-tri atk markdown 3
